Write the function separately to check token invalidity and cleanup invalid token
