+++ api/src/routers/user.js
@@ -6,14 +6,14 @@ import {
   getUserApplicationsById,
   postUserApplications,
 } from '../controllers/userController.js'
-import authenticateToken from '../middleware/token_authentication.js'
+import verifyAuthToken from '../middleware/token_authentication.js'
 
 const user = express.Router()
 
 // The endpoints for getting user information in a safe way
-user.get('/me', authenticateToken, getUserProfile)
-user.get('/applications', authenticateToken, getUserApplications)
-user.get('/applications/:id', authenticateToken, getUserApplicationsById)
-user.post('/applications', authenticateToken, postUserApplications)
+user.get('/me', verifyAuthToken, getUserProfile)
+user.get('/applications', verifyAuthToken, getUserApplications)
+user.get('/applications/:id', verifyAuthToken, getUserApplicationsById)
+user.post('/applications', verifyAuthToken, postUserApplications)
 
 export default user
